requirements.txt
openai
streamlit
python-dotenv
google-generativeai
-e .